Fix sign extension bug in EtekcityESF551Handler byte-to-uint conversion
Replace toUInt() with toUByte().toUInt() for weight and impedance byte
assembly. Byte.toUInt() sign-extends through Int, causing any byte with
bit 7 set (value >= 128) to fill upper bits with 1s before the unsigned
reinterpretation.

+++ android_app/app/src/main/java/com/health/openscale/core/bluetooth/scales/EtekcityESF551Handler.kt
@@ -109,10 +109,12 @@ class EtekcityESF551Handler : ScaleDeviceHandler() {
             return null
         }
 
-        val weightRaw = data[10].toUInt() or data[11].toUInt().shl(8) or data[12].toUInt().shl(16)
+        val weightRaw = data[10].toUByte().toUInt() or
+                data[11].toUByte().toUInt().shl(8) or
+                data[12].toUByte().toUInt().shl(16)
         val weightKg = weightRaw.toDouble() / 1000.0
-        val impedance = (data[13].toUInt() or data[14].toUInt().shl(8)).toDouble()
-//        val displayUnit = WeightUnit.fromInt(data[21].toInt())
+        val impedance = (data[13].toUByte().toUInt() or data[14].toUByte().toUInt().shl(8)).toDouble()
+        //        val displayUnit = WeightUnit.fromInt(data[21].toInt())
         val measurement = ScaleMeasurement(
             userId = user.id,
             dateTime = Date(),
